+++ test/pulsar/components/switch_test.exs
@@ -36,7 +36,7 @@ defmodule Pulsar.Components.SwitchTest do
       assert html =~ ~s(value="true")
       assert html =~ ~s(class="sr-only peer")
       assert html =~ ~s(<button type="button")
-      assert html =~ ~s(onclick="document.getElementById(&#39;)
+      assert html =~ ~s(phx-click=)
     end
 
     test "renders with field attribute" do